requirements.txt
Flask==1.1.3
google-cloud-storage
